requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
black==23.9.1
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
codespell==2.2.6
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
distlib==0.3.7
Django==5.0.4
django-vite==3.0.4
docopt==0.6.2
filelock==3.12.4
gdtoolkit==3.3.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
lark-parser==0.8.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
stripe==9.7.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
urllib3==2.2.1
virtualenv==20.24.5
zope.interface==6.4.post0
django-q2==1.6.2
